requirements.txt
google-cloud-storage==0.22.0
google-cloud-speech==0.30.0
google-cloud-datastore==1.4.0
